+++ app/selectors/accountTrackerController.ts
@@ -3,7 +3,6 @@ import {
   AccountTrackerState,
   AccountInformation,
 } from '@metamask/assets-controllers';
-import { toChecksumHexAddress } from '@metamask/controller-utils';
 import { RootState } from '../reducers';
 import { createDeepEqualSelector } from './util';
 
@@ -26,21 +25,4 @@ export const selectAccountsLength = createSelector(
   selectAccounts,
   (accounts: { [address: string]: AccountInformation }) =>
     Object.keys(accounts || {}).length,
-);
-
-export const selectAccountByChainId = createDeepEqualSelector(
-  (state: RootState) => state.engine.backgroundState,
-  (backgroundState) => {
-    const { AccountTrackerController, NetworkController, AccountsController } =
-      backgroundState;
-    const accountsByChainId = AccountTrackerController.accountsByChainId;
-    const chainId = NetworkController.providerConfig.chainId;
-    const selectedAccountId =
-      AccountsController.internalAccounts.selectedAccount;
-    const selectedAccountAddress =
-      AccountsController.internalAccounts.accounts[selectedAccountId].address;
-    return accountsByChainId[chainId][
-      toChecksumHexAddress(selectedAccountAddress)
-    ];
-  },
 );
